refactor(minicpmo): split native duplex model adapter
