+++ vllm_omni/model_executor/models/minicpmo_4_5/minicpmo_4_5_omni.py
@@ -28,7 +28,7 @@ from vllm.sequence import IntermediateTensors
 from vllm.v1.outputs import SamplerOutput
 from vllm.v1.sample.metadata import SamplingMetadata
 
-from vllm_omni.experimental.duplex.models.minicpmo45.policy import MiniCPMO45DuplexPolicy
+from vllm_omni.experimental.fullduplex.minicpmo45.policy import MiniCPMO45DuplexPolicy
 from vllm_omni.model_executor.models.minicpmo_4_5.minicpmo_4_5_omni_llm import (
     MiniCPMO45OmniLLMDummyInputsBuilder,
     MiniCPMO45OmniLLMMultiModalProcessor,
@@ -270,7 +270,7 @@ class MiniCPMO45OmniForConditionalGeneration(nn.Module, SupportsMultiModal, Supp
 
         state = helper.sessions.get(session_id)
         if state is None:
-            from vllm_omni.experimental.duplex.models.minicpmo45.runtime import (
+            from vllm_omni.experimental.fullduplex.minicpmo45.stage0 import (
                 _MiniCPMO45Stage0SessionState,
             )
 
@@ -376,7 +376,7 @@ class MiniCPMO45OmniForConditionalGeneration(nn.Module, SupportsMultiModal, Supp
         helper = getattr(self, "_minicpmo45_duplex_data_plane_helper", None)
         if helper is not None:
             return helper
-        from vllm_omni.experimental.duplex.models.minicpmo45.runtime import MiniCPMO45Stage0DuplexRuntime
+        from vllm_omni.experimental.fullduplex.minicpmo45.stage0 import MiniCPMO45Stage0DuplexRuntime
 
         model_path = getattr(getattr(self.vllm_config, "model_config", None), "model", None)
         device = str(self._module_device(self.thinker if self.thinker is not None else self))
